% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{association_statistic_for_factors}
\alias{association_statistic_for_factors}
\title{Compute the association statistic for factor variables}
\usage{
association_statistic_for_factors(
  factors,
  clean_metadata,
  na_action = "remove"
)
}
\arguments{
\item{factors}{A vector of factor variables.}

\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{na_action}{Defaults to removing rows with missing values.}
}
\description{
Returns the Cramer's V estimate and the P-value from the Pearson
chi-square score between two variables.
}
