% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{association_statistics_for_both}
\alias{association_statistics_for_both}
\title{Find intraclass correlation (ICC) between factor and continuous covariates}
\usage{
association_statistics_for_both(
  variables = names(md),
  md,
  na.action = "remove",
  alpha = 0.05
)
}
\arguments{
\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}
}
\description{
Fit a one-way ANOVA fixed effects model.
}
