% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biomart_obj}
\alias{biomart_obj}
\title{Get available Ensembl dataset}
\usage{
biomart_obj(organism, host)
}
\arguments{
\item{organism}{A character vector of the organism name. This argument takes partial strings. For example,"hsa" will match "hsapiens_gene_ensembl".}

\item{host}{An optional character vector specifying the release version. This specification is highly recommended for a reproducible workflow. (see \code{"biomaRt::listEnsemblArchives()"})}
}
\description{
Helper function to search relative Ensembl datasets by partial organism names.
}
