% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biomart_stats}
\alias{biomart_stats}
\title{Get Gene Length and GC: biomart}
\usage{
biomart_stats(gene, column_id, df, start, end)
}
\arguments{
\item{gene}{a gene name corresponding to a value in `df[,column_id]`}

\item{column_id}{the column name containg gene feature IDs}

\item{df}{the dataframe of exon/transcript annotations}

\item{start}{column name of feature start coordinate}

\item{end}{column name of feature end coordinate}
}
\description{
Get gene length and GC content of `gene` from data frame (`df`) containing
gene IDs in a column labeled the value of `column_id`. `df` must have a column
labeled sequence where the sequence info for exons/or transcripts are and
feature start and stop coordinates in columns labled the value of `start` and
`stop`. This will return a vector of `gene`, length, and GC content as a
percentage.
}
