% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boxplot_vars}
\alias{boxplot_vars}
\title{Explore metadata by variable.}
\usage{
boxplot_vars(md, include_vars, x_var)
}
\arguments{
\item{vars}{A vector of variables to visualize}
}
\value{

}
\description{
This function produces boxplots from the variables provided.
}
\examples{
TO DO: test this
}
