% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_mismatch}
\alias{check_mismatch}
\title{Check for consistent identifiers}
\usage{
check_mismatch(md, count_df)
}
\arguments{
\item{md}{A data frame with sample identifiers as row names and relevant experimental covariates.}

\item{count_df}{A counts data frame with sample identifiers as column names
and gene Ids are rownames.}
}
\description{
Subsequent steps assume unique identifiers are present within both the counts
matrix and the metadata. This function checks whether samples in the counts
matrix match samples in the metadata, and vice versa.versa.
}
