% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_covariates}
\alias{clean_covariates}
\title{Create covariate matrix from tidy metadata data frame.}
\usage{
clean_covariates(md, factors, continuous, sample_identifier)
}
\arguments{
\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{factors}{A vector of factor variables.}

\item{continuous}{A vector of continuous variables.}

\item{sample_identifier}{The name of the column with the sample identifiers that map to the gene counts data frame.}
}
\value{
A data frame with coerced variables.
}
\description{
This function takes a tidy format. Coerces vectors to correct type. Only include
covariates that have 2 or more levels. Sample identifiers are stored as rownames.
}
\examples{
data <- tibble::tribble(
 ~individualID, ~diagnosis, ~RIN,
 "ind5436", "control", 7.7,
 "ind234", "disease", 7.1
 )
clean_covariates(data, factors = c("individualID", "diagnosis"),
continuous = c("RIN"),
sample_identifier = c("individualID"))
}
