% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coerce_continuous}
\alias{coerce_continuous}
\title{Coerce objects to type numeric.}
\usage{
coerce_continuous(md, continuous)
}
\arguments{
\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{continuous}{A vector of continuous variables.}
}
\description{
Coerce objects to type numeric.
}
