% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coerce_factors}
\alias{coerce_factors}
\title{Coerce objects to type factors.}
\usage{
coerce_factors(md, factors)
}
\arguments{
\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{factors}{A vector of factor variables.}
}
\description{
Coerce objects to type factors.
}
