% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_duplicate_hgnc_symbol}
\alias{collapse_duplicate_hgnc_symbol}
\title{Duplicate HGNC}
\usage{
collapse_duplicate_hgnc_symbol(biomart_results)
}
\arguments{
\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}. Gene Ids are
stored as rownames.}
}
\description{
Count normalization requires Ensembl Ids to be unique. In rare cases, there are more
than one HGNC symbol per gene Id. This function collapses the duplicate entries into
a single entry by appending the HGNC symbols in a comma separated list.
}
