% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{complete_path}
\alias{complete_path}
\title{Create complete URL to Synapse entity}
\usage{
complete_path(synid, version = NULL)
}
\arguments{
\item{synid}{A character vector with a Synapse Id.}

\item{version}{Optional. A numeric vector with the Synapse file
version number.}
}
\description{
This function creates the url reference to entities in Synapse.
}
