% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_mean_sd}
\alias{compute_mean_sd}
\title{Compute mean and standard deviation by gene feature}
\usage{
compute_mean_sd(clean_metadata, count_df)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{count_df}{A counts data frame with sample identifiers as column names
and gene Ids are rownames.}
}
\description{
Compute mean and standard deviation by gene feature
}
