% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{conditional_plot_sexcheck}
\alias{conditional_plot_sexcheck}
\title{Conditionally wrap plot_sexcheck for drake}
\usage{
conditional_plot_sexcheck(clean_metadata, count_df, biomart_results, sex_var)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{count_df}{A counts data frame with sample identifiers as rownames.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}.}

\item{sex_var}{Column name of the sex or gender-specific metadata.}
}
\description{
Work around to expose plot_sexcheck to testing and export but also leverage
drakes function for skipping targets conditionally (see \code{"drake::cancel_if()"}).
}
