% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_geneids}
\alias{convert_geneids}
\title{Get gene Ids}
\usage{
convert_geneids(count_df)
}
\arguments{
\item{count_df}{A counts data frame with sample identifiers as column names.}
}
\description{
Get gene Ids
}
