% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{correlate_and_plot}
\alias{correlate_and_plot}
\title{Calculate correlation between covariates and principal components (PCs)}
\usage{
correlate_and_plot(
  principal_components,
  clean_metadata,
  correlation_type,
  weights = NULL,
  to_plot = FALSE,
  exclude_missing_data = NULL,
  maximum_fdr = 0.1
)
}
\arguments{
\item{principal_components}{Center, scaled and rotated principal components
matrix.}

\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{correlation_type}{Allowed values are "pearson", "spearman" and
"kendall". See \code{"psych::corr.test(method)"}.}

\item{weights}{A numeric vector of the proportion of variance explained
(PVE). Defaults to NULL.}

\item{to_plot}{Logical indicating whether correlation values should be
plotted. Defaults to FALSE.}

\item{exclude_missing_data}{A vector with column names to exclude from
`clean_metadata`.}

\item{maximum_fdr}{Maximum allowable false discovery rate (FDR). Defaults to
0.1.}
}
\value{
A list.
\itemize{
  \item plot - A customized heatmap of significant covariates and PCs
  correlated.
  \item significant_covariates - A vector of covariates where correlation
  p-value meets the FDR threshold.
  \item effects_significant_vars - A vector correlation values.
}
}
\description{
Identify covariates that significantly, as defined by a false discovery rate
(FDR) threshold, correlate with PCs computed from gene counts. This function
wraps `plot_pcs_with_covariates()` to return a heatmap to visualize this
relationship.
}
