% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cqn}
\alias{cqn}
\title{Conditional Quantile Normalization (CQN)}
\usage{
cqn(filtered_counts, biomart_results)
}
\arguments{
\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}.}
}
\description{
Normalize counts by CQN. By providing a biomart object, the systematic effect of GC content
is removed and gene length (in bp) variation is accounted for. Genes with missing GC content
or gene lengths will be removed from the counts matrix.
}
