% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dropped_genes}
\alias{dropped_genes}
\title{Dropped Genes Finder}
\usage{
dropped_genes(filtered_counts, cqn_counts)
}
\arguments{
\item{filtered_counts}{The target containing counts after low gene
expression has been removed. Defaults to target name constrained by
 \code{"targets::tar_make()"}.}

\item{cqn_counts}{A counts data frame normalized by CQN.}
}
\description{
Finds gene IDs in the filtered counts which were removed in cqn normalization
most likely as a function if no GC or length estimates being present in the
user specified biomart object.
}
