% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes}
\usage{
filter_genes(
  clean_metadata,
  count_df,
  conditions,
  cpm_threshold,
  conditions_threshold
)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{count_df}{A counts data frame with sample identifiers as column names.}

\item{conditions}{Conditions to bin gene counts that correspond to variables in `md`.}

\item{cpm_threshold}{The minimum number of CPM allowed.}

\item{conditions_threshold}{Percentage of samples that should contain the minimum CPM.}
}
\description{
Filter genes with low expression. This function is more permissive by setting conditions
that corresponds to metadata variables. The gene matrix is split by condition and the
counts per million (CPM) for a given condition is computed by \code{"sageseqr::simple_filter()"}.
}
