% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{get_association_statistics}
\alias{get_association_statistics}
\title{Visualize relationship between variables}
\usage{
get_association_statistics(md, p_value_cutoff = 0.05)
}
\arguments{
\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{p_value_cutoff}{Set the p-value threshold.}
}
\description{
Outputs a correlation matrix to visualize the relationship between
continuous and factor variables.
}
