% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Detect file type and download data}
\usage{
get_data(synid, version = NULL)
}
\arguments{
\item{synid}{A character vector with a Synapse Id.}

\item{version}{Optional. A numeric vector with the Synapse file
version number.}
}
\value{
A tibble.
}
\description{
This function takes synIds and version number to download any
rectangular file type from Synapse.
}
\examples{
\dontrun{
file <- get_data(synID = "syn1234", version = 7)

}
}
