% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gtf_stats}
\alias{gtf_stats}
\title{Get Gene Length and GC: GTF/Fasta}
\usage{
gtf_stats(i, data, genome)
}
\arguments{
\item{i}{A character of a gene id.}

\item{data}{a dataframe of exons transcrtipts or genes, imported from a GTF file.
one line per-feature, data$V1 corresponds to the featur chormosome, data$V4 to
start coordinate, data$V5 to stop coordinate, and data$V7 to the sequence strand.}

\item{genome}{a list of DNAStringSet objects named by their contig (chromosome)
name.}
}
\description{
Get gene GC content of `gene` from data frame (`df`) containing
gene IDs in a column labeled the value of `column_id`. `df` must have a column
labeled sequence where the sequence info for exons/or transcripts are and
feature start and stop coordinates in columns labled the value of `start` and
`stop`. This will return a vector of `gene`, length, and GC content as a
percentage.
}
