% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meta_express}
\alias{meta_express}
\title{Meta-analysis of continuous outcome data}
\usage{
meta_express(
  clean_metadata,
  sample_identifier,
  count_df,
  gene_id_input,
  primary_variable,
  group_variables,
  collapse_condition
)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{sample_identifier}{The name of the column with the sample identifiers that map to the gene counts data frame.}

\item{count_df}{A counts data frame with sample identifiers as column names
and gene Ids are rownames.}

\item{gene_id_input}{Column name of the gene ids in the counts_id file.}

\item{primary_variable}{The primary meta-analysis condition. The values of
the `primary_variable` will make up the experimental and control groups of
\code{"meta::metacont()"}.}

\item{group_variables}{These variables are referenced when computing summary
statistics for every gene. Therefore, a gene feature will have an observation
for each group of variables (i.e. each gene feature across unique tissue
types, sex and diagnosis).}

\item{collapse_condition}{Compute one estimate across the `collapse_condition`
which is a variabel in `clean_metadata`.}
}
\description{
Meta-analysis by standardized mean difference with the Hedges' g method
(1981).
}
