% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pairwise_meta}
\alias{pairwise_meta}
\title{Apply metacont() between two comparison groups}
\usage{
pairwise_meta(comparison_values, by_gene, primary_variable)
}
\arguments{
\item{comparison_values}{A vector of two values to compare from
`primary_variable`.}

\item{by_gene}{A grouped data frame with}

\item{primary_variable}{The primary meta-analysis condition. The values of
the `primary_variable` will make up the experimental and control groups of
\code{"meta::metacont()"}.}
}
\description{
\code{"sageseqr::meta_express()"} wraps \code{"sageseqr::pairwise_meta()"}
since the input format for `by_gene` is constrained. `by_gene` is a dataframe
grouped by gene feature and other metadata with gene summary statistics nested
under "data".
}
