% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_counts}
\alias{parse_counts}
\title{Subset counts data frame}
\usage{
parse_counts(gene_list)
}
\arguments{
\item{gene_list}{A list of genes for BioMart query.}
}
\description{
The biomaRt query requires only gene identifiers to be passed as input. Some alignment
output combines additional metadata with the counts file. This function will remove
extraneous rows and is especially meant to address output from the STAR aligner.
}
