% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_coexpression}
\alias{plot_coexpression}
\title{Co-expression histogram generation}
\usage{
plot_coexpression(cqn_counts)
}
\arguments{
\item{cqn_counts}{A counts data frame normalized by CQN.}
}
\description{
Visualize the correlation density (pairwise gene correlation) for normalized
expression data across all samples.
}
