% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_pcs_with_covariates}
\alias{plot_pcs_with_covariates}
\title{Explore relationship between principal components (PCs) and covariates}
\usage{
plot_pcs_with_covariates(
  correlation_values,
  title = NULL,
  note_threshold = NULL,
  mark_missing = NULL
)
}
\arguments{
\item{correlation_values}{A data frame that includes the p-value of the
correlation of a covariate to a PC as `pvalue`. Additionally, the data frame
should include a logical column `significant_correlations` to indicate
whether the FDR threshold is met for that variable.}

\item{title}{A string to describe the plot.}

\item{note_threshold}{A customized string to note method used to adjust
p-values for multiple comparisons. Defaults to NULL.}

\item{mark_missing}{A vector of variables to exclude from the plot due to
missing data. Defaults to NULL.}
}
\description{
Explore relationship between principal components (PCs) and covariates
}
\examples{
\dontrun{
correlation_values <- data.frame(
  compare = "PC1 (23.16\%)",
  covariates = "batch",
  pvalue = 0.56519736,
  r = "0.05729811",
  significant_correlations = FALSE
  )
note_threshold <- "FDR <= 0.1"
}
}
