% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_sexcheck_pca}
\alias{plot_sexcheck_pca}
\title{Explore metadata by gene expression on the sex chromosomes (PCA)}
\usage{
plot_sexcheck_pca(clean_metadata, count_df, biomart_results, sex_var)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{count_df}{A counts data frame with sample identifiers as column names
and gene Ids are rownames.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}. Gene Ids are
stored as rownames.}

\item{sex_var}{Column name of the sex or gender-specific metadata.}
}
\description{
Principal Component Analysis (PCA) of sex-specific genes. Samples greater
than z standard deviations (SDs) from the mean of sample sub-groups
identified as outliers.
}
