% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_results}
\alias{prepare_results}
\title{Prepare output}
\usage{
prepare_results(target, data_name, rowname = NULL)
}
\arguments{
\item{target}{The object to be stored.}

\item{data_name}{An identifier to embed in the file name.}

\item{rowname}{Optional. If applicable, the name of the variable to store
rownames.}
}
\description{
Store data in temporary files to prepare for Synapse upload.
}
