% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{provenance_helper}
\alias{provenance_helper}
\title{Provenance helper}
\usage{
provenance_helper(
  metadata_id,
  counts_id,
  metadata_version = NULL,
  counts_version = NULL,
  biomart_id = NULL,
  biomart_version = NULL
)
}
\arguments{
\item{metadata_id}{Synapse ID to clean metadata file with sample identifiers
in a column and variables of interest as column names. There cannot be any
missing values.}

\item{counts_id}{Synapse ID to counts data frame with identifiers to the
metadata as column names and gene ids in a column.}

\item{metadata_version}{Optionally, include Synapse file version number. If
omitted, current version will be downloaded.}

\item{counts_version}{Optionally, include Synapse file version number.}

\item{biomart_id}{Synapse ID to biomart object.}

\item{biomart_version}{Optionally, include Synapse file version number.}
}
\description{
Collapse Synapse Ids and version.
}
