% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rq}
\alias{Mclust}
\alias{mclustBIC}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mclust}{\code{\link[mclust]{Mclust}}, \code{\link[mclust]{mclustBIC}}}

  \item{quantreg}{\code{\link[quantreg]{rq}}}
}}

