% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{rnaseq_plan}
\alias{rnaseq_plan}
\title{Execute the drake RNA-seq plan}
\usage{
rnaseq_plan(
  metadata_id,
  metadata_version,
  counts_id,
  counts_version,
  gene_id_input,
  sample_id_input,
  factor_input,
  continuous_input,
  biomart_id,
  biomart_version,
  host,
  filters,
  organism,
  conditions,
  cpm_threshold = 1,
  conditions_threshold = 0.5,
  primary_variable,
  sex_var,
  color,
  shape,
  size,
  report_name,
  skip_model,
  de_contrasts,
  fold_change_threshold,
  p_value_threshold,
  parent_id,
  rownames,
  config_file,
  gene_list,
  force_model = NULL
)
}
\arguments{
\item{metadata_id}{Synapse ID to clean metadata file with sample identifiers
in a column and variables of interest as column names. There cannot be any
missing values.}

\item{metadata_version}{Optionally, include Synapse file version number. If
omitted, current version will be downloaded.}

\item{counts_id}{Synapse ID to counts data frame with identifiers to the
metadata as column names and gene ids in a column.}

\item{counts_version}{Optionally, include Synapse file version number.}

\item{gene_id_input}{Column name of the gene ids in the counts_id file.}

\item{sample_id_input}{Column name of the sample ids in the metadata_id file.}

\item{factor_input}{Vector of factor variables. Variables must be present
in the metadata as column names.}

\item{continuous_input}{Vector of continuous variables. Variables must be
present in the metadata as column names.}

\item{biomart_id}{Synapse ID to biomart object.}

\item{biomart_version}{Optionally, include Synapse file version number.}

\item{host}{An optional character vector specifying the release version.
This specification is highly recommended for a reproducible workflow.
(see \code{"biomaRt::listEnsemblArchives()"})}

\item{filters}{A character vector listing biomaRt query filters.
(For a list of filters see \code{"biomaRt::listFilters()"})}

\item{organism}{A character vector of the organism name. This argument
takes partial strings. For example,"hsa" will match "hsapiens_gene_ensembl".}

\item{conditions}{Optional. Conditions to bin gene counts that correspond to
variables in `md`.}

\item{cpm_threshold}{The minimum number of CPM allowed.}

\item{conditions_threshold}{Percentage of samples that should contain the minimum CPM.}

\item{primary_variable}{Baseline variable for model selection and variable to
stratify groups in the boxplot.}

\item{sex_var}{Column name of the sex or gender-specific metadata.}

\item{color}{Discrete variable in `clean_metadata` differentiated
by color.}

\item{shape}{Discrete variable in `clean_metadata` differentiated
by shape.}

\item{size}{Continuous variable in `clean_metadata` differentiated
by size.}

\item{report_name}{Name of output markdown file.}

\item{skip_model}{If TRUE, does not run regression model.}

\item{de_contrasts}{Required. Variables in the metadata to define comparisons
between groups.}

\item{fold_change_threshold}{Numeric. Significant genes are those with a log
fold-change greater than this threshold}

\item{p_value_threshold}{Numeric. P-values are adjusted by Benjamini and
Hochberg (BH) false discovery rate (FDR). Significant genes are those with an
adjusted p-value greater than this threshold.}

\item{parent_id}{A Synapse Id that corresponds to a project or
folder to store output.}

\item{rownames}{A list of variables to store rownames ordered by `metadata`,
`filtered_counts`, `biomart_results`, `cqn_counts`. If not applicable,
set as NULL.}

\item{config_file}{Optional. Path to configuration file.}

\item{gene_list}{Optional. A vector of genes to label in the volcano plot.}

\item{force_model}{Optional. A vector of variables to include in the differential
expression model if you want to skip the output of the stepwise regression.}
}
\description{
This function wraps the \code{"drake::plan()"} and copies the R markdown
report to the user's working directory.
}
