% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{run_pca}
\alias{run_pca}
\title{Principal Component Analysis (PCA) of samples}
\usage{
run_pca(normalized_counts, scaled = TRUE, percent_p_value_cutoff = 1)
}
\arguments{
\item{normalized_counts}{A counts data frame normalized by CQN, TMM, or
another preferred method, with genes as rownames.}

\item{scaled}{Defaults to TRUE. Variables scaled to have unit
variance before the analysis takes place.}

\item{percent_p_value_cutoff}{The p-value threshold in percent.}
}
\value{
A list with significant PCs rotated and PVE.
\itemize{
  \item sample_pc_values - A matrix of PCs with samples as rows.
  \item pve - A numeric vector of PVE greater than
  \code{"percent_p_value_cutoff"}.
}
}
\description{
The counts matrix is transposed to compute principal components of each
sample. The data is centered, scaled and rotated by default. The principal
components (PCs) where the proportion of variance explained (PVE) meets
the \code{"percent_p_value_cutoff"} are returned. The default percent cutoff
is 1%.
}
