% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{run_pca_and_plot_correlations}
\alias{run_pca_and_plot_correlations}
\title{Compute principal component analysis (PCA) and plot correlations}
\usage{
run_pca_and_plot_correlations(
  normalized_counts,
  clean_metadata,
  isKeyPlot = FALSE,
  scaled = TRUE,
  percent_p_value_cutoff = 1,
  correlation_type = "pearson",
  plot_covariates_vs_pca = TRUE,
  maximum_fdr = 0.1
)
}
\arguments{
\item{normalized_counts}{A counts data frame normalized by CQN, TMM, or
another preferred method, with genes as rownames.}

\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{scaled}{Defaults to TRUE. Variables scaled to have unit
variance before the analysis takes place.}

\item{percent_p_value_cutoff}{The p-value threshold in percent.}

\item{correlation_type}{Allowed values are "pearson", "spearman" and
"kendall". See \code{"psych::corr.test(method)"}.}

\item{plot_covariates_vs_pca}{Defaults to TRUE. If false, no plot is
returned.}

\item{maximum_fdr}{Maximum allowable false discovery rate (FDR). Defaults to
0.1.}
}
\value{
A list.
\itemize{
  \item significant_covariates - A vector of covariates where correlation
  p-value meets the FDR threshold.
  \item pc_results - A customized heatmap of significant covariates and PCs
  correlated.
  \item effects_significant_vars - A vector correlation values.
}
}
\description{
Identify principal components (PCs) of normalized gene counts and correlate
these PCs to interesting covariates. This function wraps
`correlate_and_plot()` to visualize, with a heatmap, the relationship between
PCs and covariates that meet a false discovery rate (FDR) threshold and
return a list of significant covariates.
}
