% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simple_filter}
\alias{simple_filter}
\title{Filter genes with low expression}
\usage{
simple_filter(count_df, cpm_threshold, conditions_threshold)
}
\arguments{
\item{cpm_threshold}{The minimum number of CPM allowed.}

\item{conditions_threshold}{Percentage of samples that should contain the minimum CPM.}
}
\description{
`simple_filter` converts a gene counts matrix into counts per million (CPM) and identifies
the genes that meet the minimum CPM threshold in a percentage of samples. The minimum CPM
threshold and percent threshold is user defined. The function returns a list of genes.
}
\examples{
\dontrun{
gene_list <- simple_filter(count_df = counts, cpm_threshold = 1, condition_threshold = 0.5)
}
}
