% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{start_de}
\alias{start_de}
\title{Initialize differential expression analysis workflow}
\usage{
start_de()
}
\description{
The `sageseqr` package provides a `targets` workflow to string together the
data processing and computational steps of RNA-seq differential expression
analysis. This funciton copies a markdown document and _targets.R file to your
working directory. The _targets.R file in your working directory is required
for the workflow to run.
}
