% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{store_results}
\alias{store_results}
\title{Store output to Synapse}
\usage{
store_results(
  parent_id,
  targets,
  names,
  inputs,
  activity_provenance,
  path_to_cache,
  rownames = NULL,
  config_file = NULL,
  report_name = NULL
)
}
\arguments{
\item{parent_id}{A Synapse Id that corresponds to a project or
folder to store output.}

\item{targets}{A list of object names to be stored.}

\item{names}{A list of human-readable names for the Synapse entities.}

\item{inputs}{A character vector of Synapse Ids to create provenance between
output files and input files.}

\item{activity_provenance}{A phrase to describe the data transformation for
provenance.}

\item{path_to_cache}{Path to Drake cache.}

\item{rownames}{A list of variables to store rownames. If not applicable,
set as NULL.}

\item{config_file}{Optional. Path to configuration file.}

\item{report_name}{Name of output markdown file and cache.}
}
\description{
Store transformed data, markdown and html report to a Folder in Synapse.
}
