% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_biotypes}
\alias{summarize_biotypes}
\title{Summarize Biotypes}
\usage{
summarize_biotypes(filtered_counts, biomart_results)
}
\arguments{
\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}. Gene Ids are
stored as rownames.}
}
\description{
Computes the fraction of genes of a particular biotype. The number of genes
must be above 100 to be summarized.
}
