% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wrap_de}
\alias{wrap_de}
\title{Wrapper for Differential Expression Analysis}
\usage{
wrap_de(
  conditions,
  filtered_counts,
  cqn_counts,
  md,
  dropped,
  biomart_results,
  p_value_threshold,
  fold_change_threshold,
  random_effect = NULL,
  model_variables = names(md),
  cores = NULL,
  cont = NULL,
  is_num = NULL,
  num_var = NULL
)
}
\arguments{
\item{conditions}{A named list of conditions to test as `primary`
in \code{"sagseqr::differential_expression()"}.}

\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{cqn_counts}{A counts data frame normalized by CQN.}

\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{dropped}{a vector of gene names to drop from filtered counts, as they
were not cqn normalized}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}. Gene Ids are
stored as rownames.}

\item{p_value_threshold}{Numeric. P-values are adjusted by Benjamini and
Hochberg (BH) false discovery rate (FDR). Significant genes are those with an
adjusted p-value greater than this threshold.}

\item{fold_change_threshold}{Numeric. Significant genes are those with a
fold-change greater than this threshold.}

\item{random_effect}{A vector of variables to consider as random effects instead
of fixed effects.}

\item{model_variables}{Optional. Vector of variables to include in the linear (mixed) model.
If not supplied, the model will include all variables in \code{md}.}

\item{cores}{An integer of cores to specify in the parallel backend (eg. 4).}

\item{cont}{Optional. A list specifying contrasts of the primary variable(s)
to consider for differential sequencing results if using factor(s) as your
primary variable. If not specified all combinations will be tested. If specified
this will speed up the pipeline. Specify the contrast with the factor values
involved in the contrast separated by a hyphen. (eg for diagnosis, \code{"primary_variable = c("AD-CT")}
where AD is the value in diagnosis column for cases and CT is the value for controls.
For multi-level contrasts, eg. \code{primary_variable = c("diagnosis", "Sex")} would have contrasts
specified as; \code{cont= c("ZZ_F-CT_F", "ZZ_M-CT_M")} to look at cases vs controls in
females and cases vs controls in males independently. While the order before or
after the hyphen doesn't matter, the order of values  before/after the underscore does matter.
The value order must be the same as the \code{"primary:} specification.
eg. \code{"primary_variable = c("diagnosis","sex")} must be \code{"CT_M-ZZ_M"} while
\code{primary_variable = c("sex","diagnosis")} must be \code{"M_CT-M_ZZ"} (default= NULL)}

\item{is_num}{Is there a numerical covariate to use as an interaction with the primary variable(s). default= NULL}

\item{num_var}{A numerical metadata column to use in an inaction with the primary variable(s). default= NULL}
}
\description{
This function will pass multiple conditions to test to \code{"sagseqr::differential_expression()"}.
}
