% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{association_statistics_for_both}
\alias{association_statistics_for_both}
\title{Find intraclass correlation (ICC) between factor and continuous covariates}
\usage{
association_statistics_for_both(
  variables = names(md),
  md,
  p_value_cutoff = 0.05,
  na_action = "remove"
)
}
\arguments{
\item{variables}{Variables to for ICC computation.Defaults to the column names
of \code{"md"}.}

\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{p_value_cutoff}{Set the p-value threshold.}

\item{na_action}{Defaults to removing rows with missing values.}
}
\description{
Fit a one-way ANOVA fixed effects model.
}
