% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{association_statistics_for_both}
\alias{association_statistics_for_both}
\title{Find intraclass correlation (ICC) between factor and continuous covariates}
\usage{
association_statistics_for_both(
  variables = names(clean_metadata),
  clean_metadata,
  p_value_cutoff = 0.05,
  na_action = "remove"
)
}
\arguments{
\item{variables}{Variables to for ICC computation.Defaults to the column names
of \code{"md"}.}

\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{p_value_cutoff}{Set the p-value threshold.}

\item{na_action}{Defaults to removing rows with missing values.}
}
\description{
Fit a one-way ANOVA fixed effects model.
}
