% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boxplot_vars}
\alias{boxplot_vars}
\title{Explore metadata by variable.}
\usage{
boxplot_vars(md, include_vars, x_var)
}
\arguments{
\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{include_vars}{A vector of variables to visualize}

\item{x_var}{Variable to plot on the x-axis.}
}
\value{

}
\description{
This function produces boxplots from the variables provided.
}
