% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_formula}
\alias{build_formula}
\title{Formula Builder}
\usage{
build_formula(md, model_variables, primary_variable)
}
\arguments{
\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{model_variables}{Vector of variables to include in the linear (mixed) model.}

\item{primary_variable}{Vector of variables that will be collapsed into a single
fixed effect interaction term.}
}
\description{
If a linear mixed model is used, all categorical variables must be
modeled as a random effect. This function identifies the variable class
to scale numeric variables and model categorical variables as a random
effect. Additionally, variables are scaled to account for
multiple variables that might have an order of magnitude difference.
}
