% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_covariates}
\alias{clean_covariates}
\title{Create covariate matrix from tidy metadata data frame.}
\usage{
clean_covariates(md, factors, continuous)
}
\value{
A data frame with coerced variables.
}
\description{
This function takes a tidy format. Coerces vectors to correct type. Only include
covariates that have 2 or more levels.
}
\examples{
data <- tibble::tribble(
 ~individualID, ~diagnosis, ~RIN,
 "ind5436", "control", 7.7,
 "ind234", "disease", 7.1
 )
clean_covariates(data, factors = c("individualID", "diagnosis"), continuous = c("RIN"))
}
