% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{differential_expression}
\alias{differential_expression}
\title{Differential Expression with Dream}
\usage{
differential_expression(
  filtered_counts,
  cqn_counts,
  md,
  primary_variable,
  biomart_results,
  model_variables = NULL
)
}
\arguments{
\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{cqn_counts}{A counts data frame normalized by CQN.}

\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{primary_variable}{Vector of variables that will be collapsed into a single
fixed effect interaction term.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}. Gene Ids are
stored as rownames.}

\item{model_variables}{Optional. Vector of variables to include in the linear (mixed) model.
If not supplied, the model will include all variables in \code{md}.}
}
\description{
Differential expression testing is performed with \code{"variancePartition::dream()"} to increase power
and decrease false positives for repeated measure designs. The `primary_variable` is modeled as a fixed
effect. If you wish to test an interaction term, `primary_variable` can take multiple variable names.
}
