% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes}
\usage{
filter_genes(md, count_df)
}
\description{
Remove genes that have less than 1 counts per million (cpm) in at least 50%
of samples per condition. If a biomaRt object is provided, gene lengths and
gene GC content is required and genes with missing values are also removed.
}
