% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes}
\usage{
filter_genes(clean_metadata, count_df, conditions)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as factors or numeric as determined
by \code{"sageseqr::clean_covariates()"})}

\item{count_df}{A counts data frame with sample identifiers as rownames.}

\item{conditions}{Conditions to bin gene counts that correspond to variables in `md`.}
}
\description{
Remove genes that have less than 1 counts per million (cpm) in at least 50%
of samples per condition.
}
