% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{get_association_statistics}
\alias{get_association_statistics}
\title{Visualize relationship between variables}
\usage{
get_association_statistics(clean_metadata, p_value_cutoff = 0.05)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{p_value_cutoff}{Set the p-value threshold.}
}
\description{
Output a correlation matrix to visualize the relationship between
continuous and factor variables. For factor variables, the Cramer's
V estimate and the P-value is computed from the Pearson chi-square
score between two variables.
}
