% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_biomart}
\alias{get_biomart}
\title{Get Ensembl biomaRt object}
\usage{
get_biomart(
  count_df,
  synid,
  version,
  host,
  filters,
  organism,
  custom,
  gtfID,
  gtfVersion = NULL,
  fastaID,
  fastaVersion = NULL,
  cores = NULL,
  isexon = FALSE
)
}
\arguments{
\item{count_df}{A counts data frame with sample identifiers as column names
and gene Ids are rownames.}

\item{synid}{A character vector with a Synapse Id.}

\item{version}{Optional. A numeric vector with the Synapse file
version number.}

\item{host}{An optional character vector specifying the release version.
This specification is highly recommended for a reproducible workflow.
(see \code{"biomaRt::listEnsemblArchives()"})}

\item{filters}{A character vector listing biomaRt query filters.
(For a list of filters see \code{"biomaRt::listFilters()"})}

\item{organism}{A character vector of the organism name. This argument
takes partial strings. For example,"hsa" will match "hsapiens_gene_ensembl".}

\item{custom}{Defaults to NULL. If TRUE, the GC and Gene Length parameters will
only consider exotic regions.}

\item{gtfID}{Defaults to NULL. A character vector with a Synapse ID
corresponding to a gtf formatted gene annotation file.}

\item{gtfVersion}{Optional. A numeric vector with the gene GTF Synapse file
version number.}

\item{fastaID}{Defaults to NULL. A character vector with a Synapse ID
corresponding to a FASTA formatted genome annotation file.}

\item{fastaVersion}{Optional. A numeric vector with the genome FASTA Synapse
file version number.}

\item{cores}{An integer of cores to specify in the parallel backend (eg. 4).}

\item{isexon}{Defaults to FALSE. If TRUE, the GC and Gene Length parameters will
only consider exotic regions.}
}
\description{
Get GC content, gene Ids, gene symbols, gene biotypes, gene lengths
and other metadata from Ensembl BioMart. Object returned contains gene Ids
as rownames.
}
