% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_biomart}
\alias{get_biomart}
\title{Get Ensembl biomaRt object}
\usage{
get_biomart(count_df, synid, version, host, filters, organism)
}
\arguments{
\item{count_df}{A counts data frame with sample identifiers as column names
and gene Ids are rownames.}

\item{synid}{Optional. A character vector with a Synapse Id.}

\item{version}{Optional. A numeric vector with the Synapse file
version number.}

\item{host}{An optional character vector specifying the release version.
This specification is highly recommended for a reproducible workflow.
(see \code{"biomaRt::listEnsemblArchives()"})}

\item{filters}{A character vector listing biomaRt query filters.
(For a list of filters see \code{"biomaRt::listFilters()"})}

\item{organism}{A character vector of the organism name. The organisms
scientific name is required to query biomaRt. This argument accepts partial
strings. For example,"hsa" will match "hsapiens_gene_ensembl" or "mmus" will
match ""mmusculus_gene_ensembl".}
}
\description{
Get GC content, gene Ids, gene symbols, gene biotypes, gene lengths
and other metadata from Ensembl BioMart. Object returned contains gene Ids
as rownames.
}
