% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Explore samples that are outliers}
\usage{
identify_outliers(
  filtered_counts,
  clean_metadata,
  color,
  shape,
  size,
  z = 4,
  split_condition = NULL
)
}
\arguments{
\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{color}{Discrete variable in `clean_metadata` differentiated
by color.}

\item{shape}{Discrete variable in `clean_metadata` differentiated
by shape.}

\item{size}{Continuous variable in `clean_metadata` differentiated
by size.}

\item{z}{Allowable number of standard deviations (SDs) from the mean.
Defaults to 4.}

\item{split_condition}{Defaults to NULL. A variable present in
`clean_metadata` is required. If provided, SDs from the mean
computed on sub-groups.}
}
\description{
Samples that are z standard deviations (SDs) from the mean of principal
components (PC) 1 and 2 are identified as outliers.
}
