% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_sexcheck}
\alias{plot_sexcheck}
\title{Explore metadata by gene expression on the sex chromosomes.}
\usage{
plot_sexcheck(clean_metadata, count_df, biomart_results, sex_var)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{count_df}{A counts data frame with sample identifiers as column names.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}.}

\item{sex_var}{Column name of the sex or gender-specific metadata.}
}
\description{
This function plots expression of X and Y marker genes, XIST and UTY respectively, and
colors each sample by the sex- or gender-specific labeling from the metadata. This is a
handy check to determine if samples were swapped or mislabeled.
}
