% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_sexcheck_pca}
\alias{plot_sexcheck_pca}
\title{Explore metadata by gene expression on the sex chromosomes (PCA)}
\usage{
plot_sexcheck_pca(clean_metadata, filtered_counts, biomart_results, sex_var)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}.}

\item{sex_var}{Column name of the sex or gender-specific metadata.}

\item{gene_annots}{Annotations that include the mapping between gene Ids and
chromosome name. Gene Ids must be rownames.}
}
\description{
Principal Component Analysis (PCA) of sex-specific genes. Samples greater
than z standard deviations (SDs) from the mean of sample sub-groups defined
}
