% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_sexcheck_pca}
\alias{plot_sexcheck_pca}
\title{Explore metadata by gene expression on the sex chromosomes (PCA)}
\usage{
plot_sexcheck_pca(
  clean_metadata,
  filtered_counts,
  gene_annots,
  sex_var,
  shape,
  size,
  z,
  split_condition
)
}
\arguments{
\item{clean_metadata}{A data frame with sample identifiers as rownames and variables as
factors or numeric as determined by \code{"sageseqr::clean_covariates()"}.}

\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{gene_annots}{Annotations that include the mapping between gene Ids and
chromosome name. Gene Ids must be rownames.}

\item{sex_var}{Column name of the sex or gender-specific metadata.}

\item{shape}{Discrete variable in `clean_metadata` differentiated
by shape.}

\item{size}{Continuous variable in `clean_metadata` differentiated
by size.}

\item{z}{Allowable number of standard deviations (SDs) from the mean.
Defaults to 4.}

\item{split_condition}{Defaults to NULL. A variable present in
`clean_metadata` is required. If provided, SDs from the mean
computed on sub-groups.}
}
\description{
Principal Component Analysis (PCA) of sex-specific genes. Samples greater
than z standard deviations (SDs) from the mean of sample sub-groups defined
by \code{"split_conditions"} are labeled in the plot.
}
