% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_results}
\alias{prepare_results}
\title{Prepare output}
\usage{
prepare_results(target, rowname = NULL)
}
\arguments{
\item{target}{The name of the object to be stored. You may pass two values
to save a target that has multidimensional arrays.}

\item{rowname}{Optional. If applicable, the name of the variable to store
rownames.}
}
\description{
Store data in temporary files to prepare for Synapse upload.
}
