% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{rnaseq_plan}
\alias{rnaseq_plan}
\title{Execute the drake RNA-seq plan}
\usage{
rnaseq_plan(
  metadata_id,
  metadata_version,
  counts_id,
  counts_version,
  gene_id_input,
  sample_id_input,
  factor_input,
  continuous_input,
  biomart_id,
  biomart_version,
  host,
  filters,
  organism,
  conditions,
  cpm_threshold = 1,
  conditions_threshold = 0.5
)
}
\arguments{
\item{metadata_id}{Synapse ID to clean metadata file with sample identifiers in a
column and variables of interest as column names. There cannot be any missing values.}

\item{metadata_version}{Optionally, include Synapse file version number. If omitted,
current version will be downloaded.}

\item{counts_id}{Synapse ID to counts data frame with identifiers to the metadata as
column names and gene ids in a column.}

\item{counts_version}{Optionally, include Synapse file version number.}

\item{gene_id_input}{Column name of the gene ids in the counts_id file.}

\item{sample_id_input}{Column name of the sample ids in the metadata_id file.}

\item{factor_input}{Vector of factor variables. Variables must be present
in the metadata as column names.}

\item{continuous_input}{Vector of continuous variables. Variables must be
present in the metadata as column names.}

\item{biomart_id}{Synapse ID to biomart object.}

\item{biomart_version}{Optionally, include Synapse file version number.}

\item{host}{An optional character vector specifying the release version.
This specification is highly recommended for a reproducible workflow.
(see \code{"biomaRt::listEnsemblArchives()"})}

\item{filters}{A character vector listing biomaRt query filters.
(For a list of filters see \code{"biomaRt::listFilters()"})}

\item{organism}{A character vector of the organism name. This argument
takes partial strings. For example,"hsa" will match "hsapiens_gene_ensembl".}

\item{conditions}{Conditions to bin gene counts that correspond to variables in `md`.}

\item{cpm_threshold}{The minimum number of CPM allowed.}

\item{conditions_threshold}{Percentage of samples that should contain the minimum CPM.}
}
\description{
This function wraps the \code{"drake::plan()"}.
}
