% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{store_results}
\alias{store_results}
\title{Store output to Synapse}
\usage{
store_results(
  clean_md = clean_md,
  filtered_counts = filtered_counts,
  biomart_results = biomart_results,
  cqn_counts = cqn_counts$counts,
  de_results,
  syn_names,
  data_names,
  parent_id,
  inputs,
  activity_provenance,
  rownames = NULL,
  config_file = NULL,
  report_name = NULL
)
}
\arguments{
\item{clean_md}{The drake target containing the metadata data frame.
Defaults to target name constrained by \code{"sageseqr::rnaseq_plan()"}.}

\item{filtered_counts}{The drake target containing counts after low gene
expression has been removed. Defaults to target name constrained by
 \code{"sageseqr::rnaseq_plan()"}.}

\item{biomart_results}{The drake target containing gene annotations from
biomart. Defaults to target name constrained by
\code{"sageseqr::rnaseq_plan()"}.}

\item{cqn_counts}{The drake target containing Conditional Quantile Normalized
(CQN) counts. Defaults to target name constrained by
\code{"sageseqr::rnaseq_plan()"}.}

\item{de_results}{The drake target containing differential expression gene
lists. Defaults to target name constrained by
\code{"sageseqr::rnaseq_plan()"}.}

\item{syn_names}{A list of human-readable names for the Synapse entities
ordered
by `metadata`, `filtered_counts`, `biomart_results`, `cqn_counts`.}

\item{data_names}{A list of identifiers to embed in the file name ordered
by `metadata`, `filtered_counts`, `biomart_results`, `cqn_counts`.}

\item{parent_id}{A Synapse Id that corresponds to a project or
folder to store output.}

\item{inputs}{A character vector of Synapse Ids to create provenance between
output files and input files.}

\item{activity_provenance}{A phrase to describe the data transformation for
provenance.}

\item{rownames}{A list of variables to store rownames ordered by `metadata`,
`filtered_counts`, `biomart_results`, `cqn_counts`. If not applicable,
set as NULL.}

\item{config_file}{Optional. Path to configuration file.}

\item{report_name}{Name of output markdown file.}
}
\description{
Store transformed data, markdown and html report to a Folder in Synapse.
}
