% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wrap_de}
\alias{wrap_de}
\title{Wrapper for Differential Expression Analysis}
\usage{
wrap_de(
  conditions,
  filtered_counts,
  cqn_counts,
  md,
  biomart_results,
  p_value_threshold,
  log_fold_threshold,
  model_variables = names(md)
)
}
\arguments{
\item{conditions}{A named list of conditions to test as `primary_variable`
in \code{"sagseqr::differential_expression()"}.}

\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{cqn_counts}{A counts data frame normalized by CQN.}

\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}. Gene Ids are
stored as rownames.}

\item{p_value_threshold}{Numeric. P-values are adjusted by Benjamini and
Hochberg (BH) false discovery rate (FDR). Significant genes are those with an
adjusted p-value greater than this threshold.}

\item{log_fold_threshold}{Numeric. Significant genes are those with a log
fold-change greater than this threshold}

\item{model_variables}{Optional. Vector of variables to include in the linear (mixed) model.
If not supplied, the model will include all variables in \code{md}.}
}
\description{
This function will pass multiple conditions to test to \code{"sagseqr::differential_expression()"}.
}
