% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wrap_de}
\alias{wrap_de}
\title{Wrapper for Differential Expression Analysis}
\usage{
wrap_de(
  conditions,
  filtered_counts,
  cqn_counts,
  md,
  model_variables,
  biomart_results
)
}
\arguments{
\item{conditions}{A list of conditions to test as `primary_variable`
in \code{"sagseqr::differential_expression()"}.}

\item{filtered_counts}{A counts data frame with genes removed that have low expression.}

\item{cqn_counts}{A counts data frame normalized by CQN.}

\item{md}{A data frame with sample identifiers in a column and relevant experimental covariates.}

\item{model_variables}{Vector of variables to include in the linear (mixed) model.}

\item{biomart_results}{Output of \code{"sageseqr::get_biomart()"}.}
}
\description{
This function will pass multiple conditions to test to \code{"sagseqr::differential_expression()"}.
}
